pyproject.toml
[project]
name = "hermione"
version = "0.1.1"
description = "A language model agent with multiple capabilities"
requires-python = ">=3.11"
dependencies = [
    "pytz",
    "langgraph",
    "langgraph-sdk",
    "langgraph-checkpoint-sqlite",
    "langsmith",
    "langchain-community",
    "langchain-core",
    "langchain-openai",
    "notebook",
    "numpy",
    "langgraph-cli",
    "grandalf",
    "fastapi",
    "uvicorn",
    "pydantic",
    "python-dotenv",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov",
    "black",
    "isort",
    "mypy",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
markers = [
    "integration: mark test as an integration test",
    "slow: mark test as slow running",
]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=src"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3 